Rename TransactionHistoryCard to PointsHistoryCard
Renamed the TransactionHistoryCard component and updated its usage in UserDashboardPage.vue to improve naming consistency and better reflect its purpose.
